+++ test/temmlPostProcess.js
@@ -14,7 +14,7 @@
    * https://mit-license.org/
    */
 
-  const version = "0.10.22";
+  const version = "0.10.23";
 
   function postProcess(block) {
     const labelMap = {};